requirements.txt
streamlit
watchdog
pymongo[srv]
pandas
MonsterLab
python-dotenv
certifi
